chore(typescript): Migrate @feathersjs/errors to modern classes and TypeScript
