+++ .gitignore
@@ -46,5 +46,6 @@ packages/authentication-client/lib
 packages/authentication-oauth/lib
 packages/configuration/lib
 packages/commons/lib
+packages/errors/lib
 packages/tests/lib
 packages/transport-commons/lib
