+++ packages/errors/.npmignore
@@ -1,1 +1,2 @@
 test/
+tsconfig.json
