+++ packages/errors/package.json
@@ -3,8 +3,8 @@
 	"description": "Common error types for Feathers apps",
 	"version": "4.5.2",
 	"homepage": "https://feathersjs.com",
-	"main": "lib/index",
-	"types": "index.d.ts",
+	"main": "lib/",
+	"types": "lib/",
 	"keywords": [
 		"feathers",
 		"feathers-plugin"
@@ -30,18 +30,22 @@
 		"lib": "lib"
 	},
 	"scripts": {
-		"test": "mocha --config ../../.mocharc.json"
+		"prepublish": "npm run compile",
+    "compile": "shx rm -rf lib/ && tsc",
+    "test": "mocha --config ../../.mocharc.ts.json --recursive test/**.test.ts test/**/*.test.ts"
 	},
 	"publishConfig": {
 		"access": "public"
 	},
-	"dependencies": {
-		"debug": "^4.1.1"
-	},
+	"dependencies": {},
 	"devDependencies": {
 		"@feathersjs/feathers": "^4.5.2",
-		"express": "^4.17.1",
-		"mocha": "^7.1.1"
+		"@types/mocha": "^7.0.2",
+    "@types/node": "^13.11.1",
+    "mocha": "^7.1.1",
+    "shx": "^0.3.2",
+    "ts-node": "^8.8.2",
+    "typescript": "^3.8.3"
 	},
 	"gitHead": "9b9f0f13387341bdd320f1e66feda828fca2c9f2"
 }
